Change ConfigIniEnv to require an optional requirement
This fixes Everett so that configobj is no longer a hard requirement
since it's only used by ConfigIniEnv and if you don't use that, then
you shouldn't need to install configobj.

Fixes #79
