+++ everett/manager.py
@@ -15,8 +15,6 @@ import re
 import sys
 import types
 
-from configobj import ConfigObj
-
 from everett import (
     ConfigurationError,
     ConfigurationMissingError,
@@ -28,6 +26,8 @@ from everett import (
 
 # This is a stack of overrides to be examined in reverse order
 _CONFIG_OVERRIDE = []
+
+# Regex for valid keys in an env file
 ENV_KEY_RE = re.compile(r'^[a-z][a-z0-9_]*$', flags=re.IGNORECASE)
 
 
@@ -50,7 +50,7 @@ def qualname(thing):
     if mod and mod.__name__ not in ('__main__', '__builtin__', 'builtins'):
         parts.append(mod.__name__)
 
-    # Python 3 has __qualname__--just use that if it's available
+    # If there's a __qualname__, use that
     if hasattr(thing, '__qualname__'):
         parts.append(thing.__qualname__)
         return '.'.join(parts)
@@ -496,131 +496,6 @@ class ConfigOSEnv(object):
     """
     def get(self, key, namespace=None):
         return get_key_from_envs(os.environ, key, namespace)
-
-
-class ConfigIniEnv(object):
-    """Source for pulling configuration from INI files
-
-    Takes a path or list of possible paths to look for the INI file and uses
-    the first one it finds.
-
-    If it finds no INI files in the possible paths, then this configuration
-    source will be a no-op.
-
-    This will expand ``~`` as well as work relative to the current working
-    directory.
-
-    This example looks just for the INI file specified in the environment::
-
-        from everett.manager import ConfigIniEnv, ConfigManager
-
-        config = ConfigManager([
-            ConfigIniEnv(os.environ.get('FOO_INI'))
-        ])
-
-
-    If there's no ``FOO_INI`` in the environment, then the path will be
-    ignored.
-
-    Here's an example that looks for the INI file specified in the environment
-    variable ``FOO_INI`` and failing that will look for ``.antenna.ini`` in the
-    user's home directory::
-
-        from everett.manager import ConfigIniEnv, ConfigManager
-
-        config = ConfigManager([
-            ConfigIniEnv([
-                os.environ.get('FOO_INI'),
-                '~/.antenna.ini'
-            ])
-        ])
-
-
-    This example looks for a ``config/local.ini`` file which overrides values
-    in a ``config/base.ini`` file both are relative to the current working
-    directory::
-
-        from everett.manager import ConfigIniEnv, ConfigManager
-
-        config = ConfigManager([
-            ConfigIniEnv('config/local.ini'),
-            ConfigIniEnv('config/base.ini')
-        ])
-
-
-    Note how you can have multiple ``ConfigIniEnv`` files and this is how you
-    can set Everett up to have values in one INI file override values in
-    another INI file.
-
-    INI files must have a "main" section. This is where keys that aren't in a
-    namespace are placed.
-
-    Minimal INI file::
-
-        [main]
-
-
-    In the INI file, namespace is a section. So key "user" in namespace "foo"
-    is::
-
-        [foo]
-        user=someval
-
-
-    Everett uses configobj, so it supports nested sections like this::
-
-        [main]
-        foo=bar
-
-        [namespace]
-        foo2=bar2
-
-          [[namespace2]]
-          foo3=bar3
-
-
-    Which gives you these:
-
-    * ``FOO``
-    * ``NAMESPACE_FOO2``
-    * ``NAMESPACE_NAMESPACE2_FOO3``
-
-    See more details here:
-    http://configobj.readthedocs.io/en/latest/configobj.html#the-config-file-format
-
-    """
-    def __init__(self, possible_paths):
-        self.cfg = {}
-        possible_paths = listify(possible_paths)
-
-        for path in possible_paths:
-            if not path:
-                continue
-
-            path = os.path.abspath(os.path.expanduser(path.strip()))
-            if path and os.path.isfile(path):
-                self.cfg.update(self.parse_ini_file(path))
-                break
-
-    def parse_ini_file(self, path):
-        cfgobj = ConfigObj(path, list_values=False)
-
-        def extract_section(namespace, d):
-            cfg = {}
-            for key, val in d.items():
-                if isinstance(d[key], dict):
-                    cfg.update(extract_section(namespace + [key], d[key]))
-                else:
-                    cfg['_'.join(namespace + [key]).upper()] = val
-
-            return cfg
-
-        return extract_section([], cfgobj.dict())
-
-    def get(self, key, namespace=None):
-        # The "main" section is considered the root mainspace.
-        namespace = namespace or ['main']
-        return get_key_from_envs(self.cfg, key, namespace)
 
 
 class ConfigManagerBase(object):
