+++ requirements-dev.txt
@@ -1,4 +1,4 @@
--e .
+-e .[ini]
 
 pytest==4.0.2
 pytest-wholenodeid==0.2
