+++ tests/test_manager.py
@@ -16,7 +16,6 @@ from everett import (
 from everett.manager import (
     ConfigDictEnv,
     ConfigEnvFileEnv,
-    ConfigIniEnv,
     ConfigManager,
     ConfigOSEnv,
     ConfigObjEnv,
@@ -279,36 +278,6 @@ def test_ConfigOSEnv():
     assert cose.get('everett_test_foo') == 'bar'
     assert cose.get('EVERETT_test_foo') == 'bar'
     assert cose.get('foo', namespace=['everett', 'test']) == 'bar'
-
-
-def test_ConfigIniEnv(datadir):
-    ini_filename = os.path.join(datadir, 'config_test.ini')
-    cie = ConfigIniEnv([ini_filename])
-    assert cie.get('foo') == 'bar'
-    assert cie.get('FOO') == 'bar'
-    assert cie.get('foo', namespace='nsbaz') == 'bat'
-    assert cie.get('foo', namespace=['nsbaz']) == 'bat'
-    assert cie.get('foo', namespace=['nsbaz', 'nsbaz2']) == 'bat2'
-
-    cie = ConfigIniEnv(['/a/b/c/bogus/filename'])
-    assert cie.get('foo') == NO_VALUE
-
-
-def test_ConfigIniEnv_multiple_files(datadir):
-    ini_filename = os.path.join(datadir, 'config_test.ini')
-    ini_filename_original = os.path.join(datadir, 'config_test_original.ini')
-    cie = ConfigIniEnv([ini_filename, ini_filename_original])
-    # Only the first found file is loaded, so foo_original does not exist
-    assert cie.get('foo_original') == NO_VALUE
-    cie = ConfigIniEnv([ini_filename_original])
-    # ... but it is there if only the original is loaded (safety check)
-    assert cie.get('foo_original') == 'original'
-
-
-def test_ConfigIniEnv_does_not_parse_lists(datadir):
-    ini_filename = os.path.join(datadir, 'config_test.ini')
-    cie = ConfigIniEnv([ini_filename])
-    assert cie.get('bar') == 'test1,test2'
 
 
 def test_ConfigEnvFileEnv(datadir):
